Squashed commit of the following:
commit 5a08893ffdc4dc96237021a84a9b52fb59f96904
Author: c08oprkiua <85906487+c08oprkiua@users.noreply.github.com>
Date:   Sun Jan 18 22:01:45 2026 -0700

    Removed SetPlayerAnimation

    It was only used in one spot, and caused a null check to appear in every case of a switch statement instead of just checking it once at the start

commit d6fee54fcc2ef5dacf902dd949bdbf1ad86183f7
Author: Dubnium32x <dbkleven@gmail.com>
Date:   Sat Jan 17 20:46:30 2026 -0800

    Revert "some fixes"

    This reverts commit 4f6168f172b5a0f6b199507f285e02e390a980ae.

commit 4f6168f172b5a0f6b199507f285e02e390a980ae
Author: Dubnium32x <dbkleven@gmail.com>
Date:   Fri Jan 16 12:26:39 2026 -0800

    some fixes

commit f95446e4fc4a55db7324f63187fabedaf922a6a7
Author: Dubnium32x <dbkleven@gmail.com>
Date:   Fri Jan 16 11:55:29 2026 -0800

    Reapply "major breakthorugh!"

    This reverts commit 94d6e146ac0771e79b3222dfe46b7b8ce9208775.

commit 94d6e146ac0771e79b3222dfe46b7b8ce9208775
Author: Dubnium32x <dbkleven@gmail.com>
Date:   Fri Jan 16 04:02:14 2026 -0800

    Revert "major breakthorugh!"

    This reverts commit 7e7fef78cf9a8682ba0ccfbf7841f97fd5984431.

commit 7e7fef78cf9a8682ba0ccfbf7841f97fd5984431
Author: Dubnium32x <dbkleven@gmail.com>
Date:   Thu Jan 15 21:14:59 2026 -0800

    major breakthorugh!

+++ SOURCE/entity/player/player-var.h
@@ -125,15 +125,15 @@ static const uint8_t ANGLE_DOWN_LEFT = 96;
 static const uint8_t ANGLE_UP_RIGHT = 224;
 static const uint8_t ANGLE_UP_LEFT = 160;
 
-// Slope angle constants
-static const uint8_t SLIP_ANGLE_START = 46;
-static const uint8_t SLIP_ANGLE_END = 315;
+// Slope angle constants (in degrees, for comparison with converted angles)
+static const uint16_t SLIP_ANGLE_START = 46;
+static const uint16_t SLIP_ANGLE_END = 315;
 
 // Alternate slope angle ranges (Sonic 3)
-static const uint8_t SLIP_ANGLE_START_S3 = 35;
-static const uint8_t SLIP_ANGLE_END_S3 = 326;
-static const uint8_t FALL_ANGLE_START_S3 = 75;
-static const uint8_t FALL_ANGLE_END_S3 = 286;
+static const uint16_t SLIP_ANGLE_START_S3 = 35;
+static const uint16_t SLIP_ANGLE_END_S3 = 326;
+static const uint16_t FALL_ANGLE_START_S3 = 75;
+static const uint16_t FALL_ANGLE_END_S3 = 286;
 
 // Spindash constants
 #define SPINDASH_BASE_SPEED 6.0f
